Performance critical code analysis actions: 1. renaming 2. message string
